[TASK] Move configuration to EM in TYPO3 6.2
Configuration was originally part of the Extension Manager.
But due to major refactoring, it was no more possible to keep
it in latest TYPO3 versions.

Since TYPO3 6.2 is now integrating a signal/slot mechanism to
extend the list of available actions on an extension, the
Backend module which was introduced to work around the limitation
should be removed and the configuration part integrated back
"into" the Extension Manager.

Change-Id: I1292de9f374932f993a24da389f0a23171f874df
Resolves: #56896
Reviewed-on: https://review.typo3.org/28379
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers