[BUGFIX] Move configuration to dedicated file
Since TYPO3 v9 it is not possible anymore to store the configuration
into LocalConfiguration.php because the hack is "detected" and the
corresponding configuration may be dropped without notice.

In order to prevent that, the configuration is now stored to a
dedicated configuration file typo3conf/image_autoresize.config.php.

The extension will automatically migrate configuration from
LocalConfiguration.php (if present) to this dedicated file.

Resolves: #18