[BUGFIX] Prevent Warning caused by invalid enum cast
In the BatchResize command, a callback method receiving a $severity variable does not correctly check the type
and tries to cast it to an int. However, the value is an enum for Typo3 >= 12 which is not castable to an int.

This Commit replaces the typecast with an access to the enums value, which is an int.

Resolves: #108